+++ src/main/java/lotto/domain/WinningLottoNumbers.java
@@ -12,28 +12,28 @@ public class WinningLottoNumbers {
         this.winningLottoNumbers = winningLottoNumbers;
     }
 
-    public static WinningLottoNumbers from(final List<Long> numbers) {
+    public static WinningLottoNumbers from(final List<Integer> numbers) {
         validate(numbers);
         return new WinningLottoNumbers(makeWinningLottoNumbers(numbers));
     }
 
-    private static void validate(final List<Long> numbers) {
+    private static void validate(final List<Integer> numbers) {
         isEachNumberUnique(numbers);
     }
 
-    private static void isEachNumberUnique(final List<Long> numbers) {
+    private static void isEachNumberUnique(final List<Integer> numbers) {
         if (isSizeLessThanSix(numbers)) {
             throw LottoInputException.of(LottoInputExceptionMessage.WINNING_LOTTO_NUMBERS_ARE_DUPLICATED);
         }
     }
 
-    private static boolean isSizeLessThanSix(final List<Long> numbers) {
+    private static boolean isSizeLessThanSix(final List<Integer> numbers) {
         return numbers.stream()
                 .distinct()
                 .count() < LottoConstant.LOTTO_ITEM_COUNT.getValue();
     }
 
-    private static List<LottoNumber> makeWinningLottoNumbers(List<Long> numbers) {
+    private static List<LottoNumber> makeWinningLottoNumbers(List<Integer> numbers) {
         return numbers.stream()
                 .map(LottoNumber::from)
                 .toList();
